backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
psycopg2-binary==2.9.11
supabase==2.25.1
python-multipart==0.0.20
pydantic==2.12.5

